+++ src/main/java/com/example/playerandgames/models/Player.java
@@ -14,7 +14,7 @@ public class Player {
 
   private String province;
 
-  private Integer phone_number;
+  private Long phone_number;
 
   public Integer getPlayer_id() {
     return player_id;
@@ -64,11 +64,11 @@ public class Player {
     this.province = province;
   }
 
-  public Integer getPhone_number() {
+  public Long getPhone_number() {
     return phone_number;
   }
 
-  public void setPhone_number(Integer phone_number) {
+  public void setPhone_number(Long phone_number) {
     this.phone_number = phone_number;
   }
 }
